LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-color-adjust/inheritance-expected.txt


FAIL Property color-scheme has initial value normal assert_equals: expected "normal" but got "auto"
PASS Property color-scheme inherits
